+++ src/Partials/_variables.scss
@@ -1,0 +1,4 @@
+//COLOURS
+$shadows-small:  multiple-box-shadow(700);
+$shadows-medium: multiple-box-shadow(200);
+$shadows-big:    multiple-box-shadow(100);
