--- src/test.js
@@ -1,1 +1,0 @@
-
